move price getter under internal/pricegetter
